Added ms sql

+++ Microservices/Banking/MicroRabbitMQ.Banking.Domain/Models/Account.cs
@@ -5,6 +5,6 @@ namespace MicroRabbitMQ.Banking.Domain.Models
     {
         public int Id { get; set; }
         public string AccountType { get; set; }
-        public decimal AcountBalance { get; set; }
+        public int AcountBalance { get; set; }
     }
 }
